test: split oversized test modules
