+++ tests/user_commands/conftest.py
@@ -1,0 +1,11 @@
+import pytest
+
+from typer_bot.commands.user_commands import (
+    UserCommands,
+)
+
+
+@pytest.fixture
+async def user_commands(mock_bot, database):
+    mock_bot.db = database
+    return UserCommands(mock_bot)
